Merged PR 4: Implemented User Categories (Static) and JSON Parsing
# Categories
Implemented static categories as outlined by Customer in product requirements.
- Income
- Expenses

# JSON Parsing
Implemented functions and classes to parse a JSON Object in JS to easily manipulate the data.
```JavaScript
class Category {
    Type: string;
    Name: string;
    Subcategories: SubCategory[];
}

class SubCategory {
    Name: string;
    isStatic: boolean;
}
```
# Functional Testing
On the debug page, as a developer I want to input a sub-category and show information for any given sub-category matching that name