Merged PR 5: Prettier
I implemented Prettier into our code so that every time you save a file, it automatically formats the file. You can edit the format settings of prettier in the .prettierrc and the .prettierignore files. There is a few settings that you need to change in VS code for it to work and I have a step-by-step on how to do it on teams.