Merged PR 13: Google Authentication and user setting
This allows the user to sign into our app using Google sign in. Once signed in, they are able to click their user setting to see their name, user icon, and the sign out button. Once signed in, the menu icon changes to to the user's icon. The sign in button also disappears once signed in and reappears once the user signs out.