Merged PR 25: User login page with email and password
This now allows the user to sign in with email and password instead of Google pop up. It also creates a document for the user in Firebase that contains the user's name, email, creation date, and last sign in. It also sends the categories json so it can later be edited by the user adding their own custom sub-categories.

Related work items: #85