Merged PR 34: Transactions Search Bar and Filter
I created a search bar on the landing page that allows the user to use a case-insensitive search to find their transactions by name. I also added a filter icon in the search bar. Once clicked on, it has a modal that has the following filters:

- Type (income/expenses)
- min amount
- max amount
- date

There is also a button in the modal to clear the filters when the user wants to reset them.

Related work items: #132