Merged PR 38: Improved Sign In Experience
The sign in page now has better error messages, a confirm password field to make sure the user enters the right password when registering, a button on the password fields so that the user can see their password. I also removed the title fields from each page, changed the menu icon location to the footer where the tab buttons are located, and disabled the swipe feature so it doesn't mess up on mobile.

Related work items: #119