Merged PR 43: Budget Pie Chart
The home page now displays a pie chart that shows all of you expenses. The total of the pie chart is the sum of your income and that is shown to the left of the pie chart. To the left of the pie chart displays your remaining balance.

Related work items: #20